agent/PaperDrop_Installer/paperdrop-src/src/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
websockets>=11.0
python-escpos>=3.0
Pillow>=10.0.0
aiohttp>=3.8.0
python-multipart>=0.0.6
python-multipart
